Update game.js
Added a listen for buttons pressed from menu html file in the submenu folder. There are three listens: One for the menu to stop displaying the menu when a button pressed. One for the play button that will start the game. One for the load button which will send you to another page.